remove feed.xml feature because it apparently broke the tags
whenever i tried to generate tags with feed.xml enabled, i got an error:

== Request: /2019/06/13/how-i-solved-a-cryptic-emacs-dumping-error/index.html
       error  build/tag/test/feed.xml
uncaught throw "Template local `tag` tried to overwrite an existing context value. Please rename the key when passing to `locals`"
/Users/username/.rvm/gems/ruby-2.5.3/gems/middleman-core-4.3.4/lib/middleman-core/template_renderer.rb:140:in `throw'

so i decided to get rid of feed.xml in order to have tags working. i
think the only purpose of feed.xml is for RSS... but i don't know for
sure. it's an investigation for another time.